requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.116.1
# grpcio==1.50.0
# grpcio-tools==1.47.0
h11==0.16.0
idna==3.10
# protobuf==3.20.3
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
VUMP-gRPC-client==1.0.2
